Пересрал заниматься хернёй Давно пора

+++ check.c
@@ -1,11 +1,11 @@
 #include "check.h"
 
-short get_n(State state)
+int get_n(State state)
 {
     return state.n;
 }
 
-long int get_N(State state)
+int get_N(State state)
 {
     return state.N;
 }
@@ -15,7 +15,7 @@ int get_amps_count(State state)
     return state.amps.n;
 }
 
-Amp get_amp(State state, long int idx)
+Amp get_amp(State state, int idx)
 {
-    return state.amps.arr[(int)idx];
+    return state.amps.arr[idx];
 }
