+++ state.h
@@ -2,7 +2,7 @@
 
 typedef struct amp //Структура для хранения амплитуд
 {
-    long int idx; //индекс
+    int idx; //индекс
     complex double amplitude; //амплитуда
 } Amp;
 
@@ -13,11 +13,11 @@ typedef struct amp //Структура для хранения амплитуд
 
 typedef struct state //структура хранения состояний
 {
-    short n; //Кубиты
-    long int N; //Кол-во элементов
+    int n; //Кубиты
+    int N; //Кол-во элементов
     Amp_Vec amps; //амплитуды
 } State;
 
-void init_state(State *state, short n, long int N); //Инициализация хранилища состояний
+void init_state(State *state, int n, int N); //Инициализация хранилища состояний
 
 void get_start_amp(State *state); //Создание первой амплитуды
